requirements.txt
yfinance>=0.2.40
alpaca-py>=0.20.0
pandas>=2.0.0
numpy>=1.26.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
